ingest_data/requirements.txt
python-dotenv

qdrant_client
openai

# langchain stuff
langchain
langchain_community
langchain_openai
langchain_core
langchain-qdrant
